added seperate file for aur packages list

+++ install-bspwm.sh
@@ -40,42 +40,47 @@ function install_video_drivers() {
     if [[ $MACHINE == *"Virtualization:"* ]]
     then
         echo "Installing virtualbox drivers"
-        sudo pacman -S virtualbox-guest-utils xf86-video-fbdev
+        func_install "virtualbox-guest-utils" "sudo pacman -S virtualbox-guest-utils"
+        func_install "xf86-video-fbdev" "sudo pacman -S xf86-video-fbdev"
     else
         echo "Installing nvidia drivers and utils"
-        sudo pacman -S nvidia nvidia-utils
+        func_install "nvidia" "sudo pacman -S nvidia"
+        func_install "nvidia-utils" "sudo pacman -S nvidia-utils"        
     fi
 }
 
 
 function install_yay_aur_helper() {
-    echo "Installing yay AUR helper:"
-    cd %HOME
-    git clone https://aur.archlinux.org/yay-git.git
-    sudo chown -R ${USERNAME}:${USERNAME} ./yay-git
-    cd yay-git
-    makepkg -si
-    cd ..
-    rm -rf yay yay-git
+    if ! [[ pacman -Qi yay-git &> /dev/null ]]
+    then 
+        echo "Installing yay AUR helper:"
+        cd %HOME
+        git clone https://aur.archlinux.org/yay-git.git
+        sudo chown -R ${USERNAME}:${USERNAME} ./yay-git
+        cd yay-git
+        makepkg -si
+        cd ..
+        rm -rf yay yay-git
+    fi
 }
 
 
 function install_packages(){
-    while IFS= read -r CURRENT_LINE
+    while IFS= read -r PACKAGE
         do
-            if ! [[ "$CURRENT_LINE" == *"#"* ]]
+            if ! [[ "$PACKAGE" == *"#"* ]]
             then 
-                INSTALL_SCRIPT=$(echo "sudo pacman -S --noconfirm --needed $CURRENT_LINE ")            
-                func_install "$CURRENT_LINE" "$INSTALL_SCRIPT"
+                INSTALL_SCRIPT=$(echo "sudo pacman -S --noconfirm --needed $PACKAGE ")            
+                func_install "$PACKAGE" "$INSTALL_SCRIPT"
             fi
     done < regular_packages.txt
 
-    while IFS= read -r CURRENT_LINE
+    while IFS= read -r PACKAGE
         do
-            if ! [[ "$CURRENT_LINE" == *"#"* ]]
+            if ! [[ "$PACKAGE" == *"#"* ]]
             then 
-                INSTALL_SCRIPT=$(echo "yay -S --noconfirm --needed $CURRENT_LINE ")
-                func_install "$CURRENT_LINE" "$INSTALL_SCRIPT"        
+                INSTALL_SCRIPT=$(echo "yay -S --noconfirm --needed $PACKAGE ")
+                func_install "$PACKAGE" "$INSTALL_SCRIPT"        
             fi
     done < aur_packages.txt    
 }
@@ -86,8 +91,11 @@ function post_install_config() {
     sudo systemctl enable NetworkManager.service
 
     echo "Enabling bluetooth:"
-    sudo systemctl enable bluetooth.service
+    sudo systemctl enable bluetooth.service     
+}
 
+
+function pre_install_config() {    
     echo "Configure reflector"
     sudo reflector -c Belarus -c Poland -c Latvia -c Lithuania -c Russia -c Ukrain -a 12 -p https -p http --save /etc/pacman.d/mirrorlist
 
@@ -95,8 +103,8 @@ function post_install_config() {
     sudo systemctl enable reflector.service        
 }
 
-function reboot() {
 
+function reboot() {
     echo "====================================================="    
     echo ":: Successfully installed! ::"    
     echo "====================================================="    
@@ -104,7 +112,9 @@ function reboot() {
     reboot
 }
 
+
 welcome
+pre_install_config
 update_and_upgrade
 install_video_drivers
 install_yay_aur_helper
